DBZ-677 Renaming record state SMTs;
Keeping old names in deprecated sub-classes for the time being.
