+++ debezium-connector-mongodb/src/test/java/io/debezium/connector/mongodb/transforms/ExtractNewDocumentStateTestIT.java
@@ -22,13 +22,13 @@ import java.util.Map;
 import static org.fest.assertions.Assertions.assertThat;
 
 /**
- * Integration test for {@link UnwrapFromMongoDbEnvelope}. It sends operations into
+ * Integration test for {@link ExtractNewDocumentState}. It sends operations into
  * MongoDB and listens on messages that are generated by Debezium plug-in. The messages
  * are then run through the SMT itself.
  *
  * @author Jiri Pechanec
  */
-public class UnwrapFromMongoDbEnvelopeTestIT extends AbstractUnwrapFromMongoDbEnvelopeTestIT {
+public class ExtractNewDocumentStateTestIT extends AbstractExtractNewDocumentStateTestIT {
 
     private static final String CONFIG_DROP_TOMBSTONES = "drop.tombstones";
     private static final String HANDLE_DELETES = "delete.handling.mode";
