+++ debezium-connector-mongodb/src/test/java/io/debezium/connector/mongodb/transforms/UpdateOperators/AbstractExtractNewDocumentStateUpdateOperatorsTestIT.java
@@ -6,8 +6,8 @@
 package io.debezium.connector.mongodb.transforms.UpdateOperators;
 
 import com.mongodb.MongoClient;
-import io.debezium.connector.mongodb.transforms.AbstractUnwrapFromMongoDbEnvelopeTestIT;
-import io.debezium.connector.mongodb.transforms.UnwrapFromMongoDbEnvelope;
+import io.debezium.connector.mongodb.transforms.AbstractExtractNewDocumentStateTestIT;
+import io.debezium.connector.mongodb.transforms.ExtractNewDocumentState;
 import org.apache.kafka.connect.source.SourceRecord;
 import org.bson.Document;
 
@@ -16,7 +16,7 @@ import java.util.function.Consumer;
 import static org.fest.assertions.Assertions.assertThat;
 
 /**
- * Integration test for {@link UnwrapFromMongoDbEnvelope}. It sends operations into
+ * Integration test for {@link ExtractNewDocumentState}. It sends operations into
  * MongoDB and listens on messages that are generated by Debezium plug-in. The messages
  * are then run through the SMT itself.
  * <p>
@@ -25,7 +25,7 @@ import static org.fest.assertions.Assertions.assertThat;
  *
  * @author Renato Mefi
  */
-abstract class AbstractUnwrapFromMongoDbEnvelopeUpdateOperatorsTestIT extends AbstractUnwrapFromMongoDbEnvelopeTestIT {
+abstract class AbstractExtractNewDocumentStateUpdateOperatorsTestIT extends AbstractExtractNewDocumentStateTestIT {
 
     @Override
     protected String getCollectionName() {
