+++ debezium-core/src/test/java/io/debezium/transforms/ExtractNewRecordStateTest.java
@@ -24,7 +24,7 @@ import io.debezium.doc.FixFor;
 /**
  * @author Jiri Pechanec
  */
-public class UnwrapFromEnvelopeTest {
+public class ExtractNewRecordStateTest {
 
     private static final String DROP_TOMBSTONES = "drop.tombstones";
     private static final String HANDLE_DELETES = "delete.handling.mode";
@@ -32,7 +32,7 @@ public class UnwrapFromEnvelopeTest {
 
     @Test
     public void testTombstoneDroppedByDefault() {
-        try (final UnwrapFromEnvelope<SourceRecord> transform = new UnwrapFromEnvelope<>()) {
+        try (final ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
             transform.configure(props);
 
@@ -43,7 +43,7 @@ public class UnwrapFromEnvelopeTest {
 
     @Test
     public void testTombstoneDroppedConfigured() {
-        try (final UnwrapFromEnvelope<SourceRecord> transform = new UnwrapFromEnvelope<>()) {
+        try (final ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
             props.put(DROP_TOMBSTONES, "true");
             transform.configure(props);
@@ -55,7 +55,7 @@ public class UnwrapFromEnvelopeTest {
 
     @Test
     public void testTombstoneForwardConfigured() {
-        try (final UnwrapFromEnvelope<SourceRecord> transform = new UnwrapFromEnvelope<>()) {
+        try (final ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
             props.put(DROP_TOMBSTONES, "false");
             transform.configure(props);
@@ -120,7 +120,7 @@ public class UnwrapFromEnvelopeTest {
 
     @Test
     public void testDeleteDroppedByDefault() {
-        try (final UnwrapFromEnvelope<SourceRecord> transform = new UnwrapFromEnvelope<>()) {
+        try (final ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
             transform.configure(props);
 
@@ -131,7 +131,7 @@ public class UnwrapFromEnvelopeTest {
 
     @Test
     public void testHandleDeleteDrop() {
-        try (final UnwrapFromEnvelope<SourceRecord> transform = new UnwrapFromEnvelope<>()) {
+        try (final ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
             props.put(HANDLE_DELETES, "drop");
             transform.configure(props);
@@ -143,7 +143,7 @@ public class UnwrapFromEnvelopeTest {
 
     @Test
     public void testHandleDeleteNone() {
-        try (final UnwrapFromEnvelope<SourceRecord> transform = new UnwrapFromEnvelope<>()) {
+        try (final ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
             props.put(HANDLE_DELETES, "none");
             transform.configure(props);
@@ -156,7 +156,7 @@ public class UnwrapFromEnvelopeTest {
 
     @Test
     public void testHandleDeleteRewrite() {
-        try (final UnwrapFromEnvelope<SourceRecord> transform = new UnwrapFromEnvelope<>()) {
+        try (final ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
             props.put(HANDLE_DELETES, "rewrite");
             transform.configure(props);
@@ -169,7 +169,7 @@ public class UnwrapFromEnvelopeTest {
 
     @Test
     public void testHandleCreateRewrite() {
-        try (final UnwrapFromEnvelope<SourceRecord> transform = new UnwrapFromEnvelope<>()) {
+        try (final ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
             props.put(HANDLE_DELETES, "rewrite");
             props.put(OPERATION_HEADER, "true");
@@ -179,14 +179,14 @@ public class UnwrapFromEnvelopeTest {
             final SourceRecord unwrapped = transform.apply(createRecord);
             assertThat(((Struct) unwrapped.value()).getString("__deleted")).isEqualTo("false");
             assertThat(unwrapped.headers()).hasSize(1);
-            String headerValue = getSourceRecordHeaderByKey(unwrapped, UnwrapFromEnvelope.DEBEZIUM_OPERATION_HEADER_KEY);
+            String headerValue = getSourceRecordHeaderByKey(unwrapped, ExtractNewRecordState.DEBEZIUM_OPERATION_HEADER_KEY);
             assertThat(headerValue).isEqualTo(Envelope.Operation.CREATE.code());
         }
     }
 
     @Test
     public void testUnwrapCreateRecord() {
-        try (final UnwrapFromEnvelope<SourceRecord> transform = new UnwrapFromEnvelope<>()) {
+        try (final ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
             transform.configure(props);
 
@@ -198,7 +198,7 @@ public class UnwrapFromEnvelopeTest {
 
     @Test
     public void testIgnoreUnknownRecord() {
-        try (final UnwrapFromEnvelope<SourceRecord> transform = new UnwrapFromEnvelope<>()) {
+        try (final ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
             transform.configure(props);
 
@@ -213,7 +213,7 @@ public class UnwrapFromEnvelopeTest {
     @Test
     @FixFor("DBZ-971")
     public void testUnwrapPropagatesRecordHeaders() {
-        try (final UnwrapFromEnvelope<SourceRecord> transform = new UnwrapFromEnvelope<>()) {
+        try (final ExtractNewRecordState<SourceRecord> transform = new ExtractNewRecordState<>()) {
             final Map<String, String> props = new HashMap<>();
             transform.configure(props);
 
@@ -227,6 +227,19 @@ public class UnwrapFromEnvelopeTest {
             Iterator<Header> headers = unwrapped.headers().allWithName("application/debezium-test-header");
             assertThat(headers.hasNext()).isTrue();
             assertThat(headers.next().value().toString()).isEqualTo("shouldPropagatePreviousRecordHeaders");
+        }
+    }
+
+    @Test
+    @FixFor("DBZ-677")
+    public void canUseDeprecatedSmt() {
+        try (final UnwrapFromEnvelope<SourceRecord> transform = new UnwrapFromEnvelope<>()) {
+            final Map<String, String> props = new HashMap<>();
+            transform.configure(props);
+
+            final SourceRecord createRecord = createCreateRecord();
+            final SourceRecord unwrapped = transform.apply(createRecord);
+            assertThat(((Struct) unwrapped.value()).getInt8("id")).isEqualTo((byte) 1);
         }
     }
 }
